refactor(fiber): make fiber a function that can halt and resume
